reorgnize the project
