+++ .env
@@ -1,0 +1,3 @@
+URL_MONGO = 'mongodb://127.0.0.1/blog'
+ACCESS_TOKEN = "access toke"
+REFRESH_TOKEN = "refresh toekn"
