+++ models/post.js
@@ -18,13 +18,11 @@ const postSchema = new mongoose.Schema({
    user:{
       type:mongoose.Types.ObjectId,
       ref:'User'
+   },
+   image:{
+      type:String,
    }
 })
-
-
-
-
-
 
 module.exports = mongoose.model("Post",postSchema)
