+++ server.js
@@ -2,23 +2,18 @@ require('dotenv').config();
 const express = require('express')
 const cookies = require('cookie-parser');
 const connection = require('./configurations/mongoodb');
+const userRouter = require('./routers/userRouter');
+const postRouter = require('./routers/postRouter');
 const mongoose = require('mongoose')
-const auth = require('./middlewares/auth')
-const role = require('./middlewares/role')
+
 const app = express()
 connection();
 app.use(express.json())
 app.use(express.urlencoded({ extended: true }));
 app.use(cookies())
 const port = 3000 || process.env.PORT ;
-
-app.use('/Blog',require('./routers/get'))
-app.use('/Blog/regester',require('./routers/newUser'))
-app.use('/Blog/sing',require('./routers/singIn'))
-app.use(auth)
-app.use('/Blog/new/post',require('./routers/newPost'))
-app.use('/Blog/delete',role,require('./routers/delete'))
-app.use('/Blog/update',require('./routers/update'))
+app.use('/Blog/user',userRouter)
+app.use('Blog/post',postRouter)
 
 mongoose.connection.once('open',()=>{
    console.log('connect to database')
